requirements.txt
groq
python-dotenv
flask
streamlit
gradio